% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_properties.R
\name{add_properties}
\alias{add_properties}
\title{Add properties}
\usage{
add_properties(x, ..., after = length(x))
}
\arguments{
\item{x}{A list or vector.}

\item{...}{Elements to add, as named arguments.}

\item{after}{Position after which elements should be added.
Use \code{0} to add elements at the beginning.}
}
\value{
\code{x} with the added elements.
}
\description{
Adds or inserts provided arguments as elements to a list.
Attributes and classes of the original list are retained.
}
\details{
\code{add_properties()} can be used to add (custom) metadata properties to a Data
Package, Data Resource, Table Dialect or Table Schema.
Note that added properties are not validated.
See the vignettes for an overview of all standard metadata properties.

\code{add_properties()} has some advantages over:
\itemize{
\item \code{\link[=append]{append()}}: removes attributes and classes, invalidating a Data Package
object.
\item Direct assignment (\code{package$title <- "My package"}): no control over the
position of the added element.
}

\code{add_resource()} also supports adding metadata properties to a Data Resource,
except those automatically set by that function.
}
\examples{
# Add property to a generic list
list <- list(a = 1, b = 2)
add_properties(list, added = 3)

# Add properties to a Data Package
package <- create_package()
package <- add_properties(
  package,
  title = "Example package",
  keywords = c("camera traps", "frictionlessdata"),
  after = 1 # Add after the first property
)
package

str(package)
}
\seealso{
Other edit functions:
\code{\link[=add_resource]{add_resource()}},
\code{\link[=remove_resource]{remove_resource()}}
}
\concept{edit functions}
