% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.datapackage.R
\name{append.datapackage}
\alias{append.datapackage}
\title{Append a Data Package}
\usage{
\method{append}{datapackage}(x, values, after = length(x))
}
\arguments{
\item{x}{Data Package object, as returned by \code{\link[=read_package]{read_package()}} or
\code{\link[=create_package]{create_package()}}.}

\item{values}{to be included in the modified vector.}

\item{after}{a subscript, after which the values are to be appended.}
}
\value{
A Data Package object with the elements of \code{values} appended after
the specified element of \code{x}.
}
\description{
Appends values to a Data Package object, preserving its attributes such as
directory and class.
}
\examples{
p <- example_package()
p_app <- append(p, list("$schema" = "just added"), after = 0)
p_app
attributes(p_app)
}
