% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.datapackage.R
\name{append}
\alias{append}
\title{Append values to an object}
\usage{
append(x, values, after = length(x))
}
\arguments{
\item{x}{the vector the values are to be appended to.}

\item{values}{to be included in the modified vector.}

\item{after}{a subscript, after which the values are to be appended.}
}
\value{
A vector containing the values in \code{x} with the elements of \code{values}
appended after the specified element of \code{x}.
}
\description{
An S3 generic version of \code{\link[base:append]{base::append()}}; see \code{\link[=append.datapackage]{append.datapackage()}}.
}
