% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.list.datapackage}
\alias{as.list.datapackage}
\title{Convert a data package to a bare list}
\usage{
\method{as.list}{datapackage}(x)
}
\arguments{
\item{x}{Data package to convert.}
}
\value{
The data package converted to a bare list
}
\description{
Convert a data package to a bare list
}
\concept{helper functions}
