% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_package.R
\name{check_package}
\alias{check_package}
\title{Check a Data Package object}
\usage{
check_package(package)
}
\value{
\code{TRUE} or error.
}
\description{
Check if an object is a Data Package object (see \code{\link[=create_package]{create_package()}}) with
the required properties.
}
\concept{check functions}
