% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{get_schema}
\alias{resources}
\title{Deprecated functions}
\usage{
get_schema(package, resource_name)

resources(package)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use \code{\link[=schema]{schema()}} instead of \code{get_schema()}.

Use \code{\link[=resource_names]{resource_names()}} instead of \code{resources()}.
}
\concept{deprecated functions}
\keyword{internal}
