% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_package.R
\name{example_package}
\alias{example_package}
\title{Read the example Data Package}
\usage{
example_package(version = "2.0")
}
\arguments{
\item{version}{Data Package standard version number.}
}
\value{
A Data Package object, see \code{\link[=create_package]{create_package()}}.
}
\description{
Reads the example Data Package included in \code{frictionless}.
This dataset is used in examples, vignettes, and tests and contains dummy
camera trap data organized in 3 Data Resources:
\enumerate{
\item \code{deployments}: one local data file referenced in
\code{"path": "deployments.csv"}.
\item \code{observations}: two local data files referenced in
\verb{"path": ["observations_1.tsv", "observations_2.tsv"]}.
\item \code{media}: inline data stored in \code{data}.
}
}
\details{
The example Data Package is available in two versions:
\itemize{
\item \code{1.0}: specified in the \href{https://specs.frictionlessdata.io/}{v1}
specification.
\item \code{2.0}: specified in the \href{https://datapackage.org/}{v2} specification.
}
}
\examples{
# Load the v2 example Data Package (default)
example_package()

# Load the v1 example Data Package
example_package(version = "1.0")
}
\concept{sample data}
