% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frictionless-package.R
\docType{package}
\name{frictionless-package}
\alias{frictionless}
\alias{frictionless-package}
\title{frictionless: Read and Write Frictionless Data Packages}
\description{
Read and write Frictionless Data Packages. A 'Data Package' (\url{https://specs.frictionlessdata.io/data-package/}) is a simple container format and standard to describe and package a collection of (tabular) data. It is typically used to publish FAIR (\url{https://www.go-fair.org/fair-principles/}) and open datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frictionlessdata/frictionless-r}
  \item \url{https://docs.ropensci.org/frictionless/}
  \item Report bugs at \url{https://github.com/frictionlessdata/frictionless-r/issues}
}

}
\author{
\strong{Maintainer}: Peter Desmet \email{peter.desmet@inbo.be} (\href{https://orcid.org/0000-0002-8442-8025}{ORCID}) (affiliation: Research Institute for Nature and Forest (INBO))

Authors:
\itemize{
  \item Peter Desmet \email{peter.desmet@inbo.be} (\href{https://orcid.org/0000-0002-8442-8025}{ORCID}) (affiliation: Research Institute for Nature and Forest (INBO))
  \item Damiano Oldoni \email{damiano.oldoni@inbo.be} (\href{https://orcid.org/0000-0003-3445-7562}{ORCID}) (affiliation: Research Institute for Nature and Forest (INBO))
  \item Pieter Huybrechts \email{pieter.huybrechts@inbo.be} (\href{https://orcid.org/0000-0002-6658-6062}{ORCID}) (affiliation: Research Institute for Nature and Forest (INBO))
  \item Sanne Govaert \email{sanne.govaert@inbo.be} (\href{https://orcid.org/0000-0002-8939-1305}{ORCID}) (affiliation: Research Institute for Nature and Forest (INBO))
}

Other contributors:
\itemize{
  \item Kyle Husmann \email{kdh38@psu.edu} (\href{https://orcid.org/0000-0001-9875-8976}{ORCID}) (affiliation: Pennsylvania State University) [contributor]
  \item Research Institute for Nature and Forest (INBO) (\href{https://ror.org/00j54wy13}{ROR}) [copyright holder]
  \item Research Foundation - Flanders (https://lifewatch.be) [funder]
  \item Beatriz Milz \email{milz.bea@gmail.com} (\href{https://orcid.org/0000-0002-3064-4486}{ORCID}) [reviewer]
  \item João Martins \email{joao.martins@gmx.net} (\href{https://orcid.org/0000-0001-7961-4280}{ORCID}) [reviewer]
}

}
\keyword{internal}
