% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.datapackage}
\alias{print.datapackage}
\title{Print a Data Package}
\usage{
\method{print}{datapackage}(package)
}
\arguments{
\item{package}{Data Package object, created with \code{\link[=read_package]{read_package()}} or
\code{\link[=create_package]{create_package()}}.}
}
\description{
Prints a human-readable summary of a Data Package, including its resources
and a link to more information (if provided in \code{package$id}).
}
\examples{
# Read a datapackage.json file
package <- read_package(
  system.file("extdata", "datapackage.json", package = "frictionless")
)

# Print a summary of the Data Package
print(package)
}
\concept{print functions}
