% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.fls_pkg}
\alias{print.fls_pkg}
\title{Print a data package}
\usage{
\method{print}{fls_pkg}(x)
}
\arguments{
\item{x}{Data package to print}
}
\description{
Print a data package
}
\examples{
# Read a datapackage.json file
package <- read_package(
  system.file("extdata", "datapackage.json", package = "frictionless")
)

# Print out a summary of the data package
print(package)
}
\seealso{
Other read functions: 
\code{\link{read_package}()},
\code{\link{read_resource}()},
\code{\link{resources}()}
}
\concept{read functions}
