% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_resource.R
\name{remove_resource}
\alias{remove_resource}
\title{Remove a Data Resource}
\usage{
remove_resource(package, resource_name)
}
\arguments{
\item{package}{Data Package object, as returned by \code{\link[=read_package]{read_package()}} or
\code{\link[=create_package]{create_package()}}.}

\item{resource_name}{Name of the Data Resource.}
}
\value{
\code{package} with one fewer resource.
}
\description{
Removes a Data Resource from a Data Package, i.e. it removes one of the
described \code{resources}.
}
\examples{
# Load the example Data Package
package <- example_package()

# List the resources
resource_names(package)

# Remove the resource "observations"
package <- remove_resource(package, "observations")

# List the resources ("observations" removed)
resource_names(package)
}
\seealso{
Other edit functions:
\code{\link[=add_properties]{add_properties()}},
\code{\link[=add_resource]{add_resource()}}
}
\concept{edit functions}
