% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_null}
\alias{replace_null}
\title{Replace value when NULL}
\usage{
replace_null(value, replacement)
}
\arguments{
\item{value}{Value to test.}

\item{replacement}{Replacement value when \code{value} is \code{NULL}.}
}
\value{
\code{value} when not \code{NULL}, otherwise \code{replacement}.
}
\description{
Replace value when NULL
}
\keyword{internal}
