% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{resource}
\alias{resource}
\title{Get a Data Resource}
\usage{
resource(package, resource_name)
}
\arguments{
\item{package}{Data Package object, as returned by \code{\link[=read_package]{read_package()}} or
\code{\link[=create_package]{create_package()}}.}

\item{resource_name}{Name of the Data Resource.}
}
\value{
List describing a Data Resource, with new attribute \code{data_location}
to indicate how the data are attached.
If present, \code{path} will be updated to contain the full path(s).
}
\description{
Returns a Data Resource from a Data Package, i.e. the content of one of the
described \code{resources}.
}
\examples{
package <- example_package()
resource(package, "deployments")
}
\seealso{
Other accessor functions:
\code{\link[=resource_names]{resource_names()}},
\code{\link[=schema]{schema()}}
}
\concept{accessor functions}
