% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{unique_sorted}
\alias{unique_sorted}
\title{Get unique vector values sorted by how often they occur}
\usage{
unique_sorted(x)
}
\arguments{
\item{x}{Vector, e.g. \code{c("a", "b", "b", "b", "a")}.}
}
\value{
Vector with unique values sorted by occurrence, e.g. \code{c("b", "a")}.
}
\description{
Get unique vector values sorted by how often they occur
}
\keyword{interal}
