% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{version}
\alias{version}
\title{Get the specification version number}
\usage{
version(x)
}
\arguments{
\item{x}{A list describing either a Data Package, Data Resource, Table
Dialect or Table Schema.}
}
\value{
Data Package standard version number.
}
\description{
Determines what version of the Data Package, Data Resource, Table Dialect or
Table Schema standard is used, based on the \href{https://datapackage.org/standard/data-package/#dollar-schema}{\verb{$schema}} property.
\itemize{
\item \code{"1.0"}: \href{https://specs.frictionlessdata.io/}{v1} specification.
Assumed if \verb{$schema} is missing.
\item \verb{"2.0}: \href{https://datapackage.org/}{v2} specification.
\item \code{">=2.0"}: assumed if \verb{$schema} is defined by deviating from the default
(e.g. an \href{https://datapackage.org/standard/extensions/}{extension}).
}
}
\examples{
# Data Package
package <- example_package()
version(package)

# Data Resource
resource <- resource(package, "observations")
version(resource)

# Table Dialect
version(resource$dialect)

# Table Schema
schema <- schema(package, "observations")
version(schema)
}
\seealso{
Other versioning functions:
\code{\link[=upgrade_package]{upgrade_package()}}
}
\concept{versioning functions}
