% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_properties.R
\name{add_properties}
\alias{add_properties}
\title{Add properties}
\usage{
add_properties(x, ..., after = length(x))
}
\arguments{
\item{x}{A list or vector.}

\item{...}{Elements to add, as named arguments.}

\item{after}{Position after which elements should be added.
Use \code{0} to add elements at the beginning.}
}
\value{
\code{x} with the added elements.
}
\description{
Adds provided arguments as elements to a list.
Unlike \code{\link[=append]{append()}} the original attributes and classes are retained.
Can be used to add (custom) properties to a Data Package, Data Resource,
Table Dialect or Table Schema.
}
\details{
You can also use \code{\link[=add_resource]{add_resource()}} to add properties to a Data Resource.
}
\examples{
# Add property to a generic list
list <- list(a = 1, b = 2)
add_properties(list, added = 3)

# Add properties to a Data Package
package <- create_package()
(package <- add_properties(
  package,
  title = "Example package",
  keywords = c("example", "camera trap data"),
  after = 1 # After first property
))

str(package)
}
\seealso{
Other edit functions:
\code{\link[=add_resource]{add_resource()}},
\code{\link[=remove_resource]{remove_resource()}}
}
\concept{edit functions}
