% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_resource.R
\name{add_resource}
\alias{add_resource}
\title{Add a Data Resource}
\usage{
add_resource(
  package,
  resource_name,
  data,
  schema = NULL,
  replace = FALSE,
  delim = ",",
  ...
)
}
\arguments{
\item{package}{Data Package object, as returned by \code{\link[=read_package]{read_package()}} or
\code{\link[=create_package]{create_package()}}.}

\item{resource_name}{Name of the Data Resource.}

\item{data}{Data to attach, either a data frame or path(s) to CSV file(s):
\itemize{
\item Data frame: attached to the resource as \code{data} and written to a CSV file
when using \code{\link[=write_package]{write_package()}}.
\item One or more paths or URLs to CSV files as a character (vector): added to
the resource as \code{path}.
The last file will be read with \code{\link[readr:read_delim]{readr::read_delim()}} to create or
compare with \code{schema} and to set \code{format}, \code{mediatype} and \code{encoding}.
The other files are ignored, but are expected to have the same structure
and properties.
}}

\item{schema}{Either a list, or path or URL to a JSON file describing a Table
Schema for the \code{data}.
If not provided, one will be created using \code{\link[=create_schema]{create_schema()}}.}

\item{replace}{If \code{TRUE}, allows an existing resource of the same name to be
replaced.}

\item{delim}{Delimiter for the CSV file(s) referenced in \code{data} (e.g. \verb{\\t}
for a tab-separated file).
Will be set as \code{delimiter} in the resource Table Dialect, so read functions
know how to read the file(s).
Ignored if \code{data} is a data frame.}

\item{...}{Additional \href{https://datapackage.org/standard/data-resource/#properties}{metadata properties} to add to the
resource.
Note that added properties are not validated and ignored by
\code{\link[=read_resource]{read_resource()}}.
The following properties are automatically set and can't be provided with
\code{...}: \verb{$schema}, \code{name}, \code{path}, \code{data}, \code{type}, \code{format}, \code{mediatype},
\code{encoding}, \code{dialect} and \code{schema}.}
}
\value{
\code{package} with one additional resource.
}
\description{
Adds a Data Resource to a Data Package.
The resource will be a \href{https://datapackage.org/standard/data-resource/#tabular}{Tabular Data Resource}.
}
\details{
See \code{vignette("data-resource")} (and to a lesser extend
\code{vignette("table-dialect")}) to learn how this function implements the
Data Package standard.
}
\examples{
# Load the example Data Package
package <- example_package()

# List the resources
resource_names(package)

# Create a data frame
df <- data.frame(
  multimedia_id = c(
    "aed5fa71-3ed4-4284-a6ba-3550d1a4de8d",
    "da81a501-8236-4cbd-aa95-4bc4b10a05df"
  ),
  x = c(718, 748),
  y = c(860, 900)
)

# Add the resource "positions" from the data frame
package <- add_resource(package, "positions", data = df)

# Add the resource "positions_with_schema", with a user-defined schema and title
my_schema <- create_schema(df)
package <- add_resource(
  package,
  resource_name = "positions_with_schema",
  data = df,
  schema = my_schema,
  title = "Positions with schema"
)

# Replace the resource "observations" with a file-based resource (2 TSV files)
path_1 <-
  system.file("extdata", "v1", "observations_1.tsv", package = "frictionless")
path_2 <-
  system.file("extdata", "v1", "observations_2.tsv", package = "frictionless")
package <- add_resource(
  package,
  resource_name = "observations",
  data = c(path_1, path_2),
  replace = TRUE,
  delim = "\t"
)

# List the resources ("positions" and "positions_with_schema" added)
resource_names(package)
}
\seealso{
Other edit functions:
\code{\link[=add_properties]{add_properties()}},
\code{\link[=remove_resource]{remove_resource()}}
}
\concept{edit functions}
