% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_resource.R
\name{add_resource}
\alias{add_resource}
\title{Add a Data Resource}
\usage{
add_resource(package, resource_name, data, schema = NULL, ...)
}
\arguments{
\item{package}{List object describing a Data Package, created with
\code{\link[=read_package]{read_package()}} or \code{\link[=create_package]{create_package()}}.}

\item{resource_name}{Name of the Data Resource.}

\item{data}{Data to attach, either a data frame or path(s) to CSV file(s).
\itemize{
\item Data frame: attached to the resource as \code{data} and written to a CSV file
when using \code{\link[=write_package]{write_package()}}.
\item Path(s) to CSV file(s): added to the resource as \code{path}.
One file (default: last) will be read with
\link[readr:read_delim(delim = ",")]{readr::read_delim(delim = ",")} to create or compare with \code{schema}.
The other files are ignored but are expected to have the same structure.
}}

\item{schema}{List object describing a Table Schema for the \code{data}.
If not provided, one will be created using \code{\link[=create_schema]{create_schema()}}.}

\item{...}{Optional arguments passed to \code{\link[readr:read_delim]{readr::read_delim()}} when reading
path in \code{data}, e.g. \code{delim = "\\t"} for tab delimited file.}
}
\value{
Provided \code{package} with one additional resource.
}
\description{
Adds a Tabular \href{https://specs.frictionlessdata.io/data-resource/}{Data Resource} to a Data
Package.
The resource will be a \href{https://specs.frictionlessdata.io/tabular-data-resource/}{Tabular Data Resource}.
The resource name can only contain lowercase alphanumeric characters plus
\code{.}, \code{-} and \verb{_}.
}
\examples{
# Load the example Data Package
package <- example_package

# List the resource names
package$resource_names

# Create a data frame
df <- data.frame(
  multimedia_id = c(
    "aed5fa71-3ed4-4284-a6ba-3550d1a4de8d",
    "da81a501-8236-4cbd-aa95-4bc4b10a05df"
  ),
  x = c(718, 748),
  y = c(860, 900)
)

# Add a new resource "positions" to the Data Package, from the data frame
package <- add_resource(package, "positions", data = df)

# Add a new resource "deployments2" to the Data Package, from a CSV file path
path <- system.file("extdata", "deployments.csv", package = "frictionless")
package <- add_resource(package, "deployments2", data = path)

# List the resource names ("positions" and "deployments2" added)
package$resource_names
}
\seealso{
Other edit functions: 
\code{\link{get_schema}()},
\code{\link{remove_resource}()}
}
\concept{edit functions}
