% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_resource.R
\name{add_resource}
\alias{add_resource}
\title{Add a Data Resource}
\usage{
add_resource(package, resource_name, data, schema = NULL)
}
\arguments{
\item{package}{List object describing a Data Package, created with
\code{\link[=read_package]{read_package()}} or \code{\link[=create_package]{create_package()}}.}

\item{resource_name}{Name of the Data Resource.}

\item{data}{Data to attach, either a data frame or path(s) to CSV file(s).
\itemize{
\item data frame: will be attached to the resource as \code{data} and written to a
CSV file when using \code{\link[=write_package]{write_package()}}.
\item path(s) to CSV file(s): will be added to the resource as \code{path}.
The (last in case of multiple) CSV file will be read with
\code{\link[readr:read_delim]{readr::read_csv()}} using default parameters (comma-separated, headers
present, etc.) only for the purpose of creating or comparing with a
\code{schema}.
}}

\item{schema}{List object describing a Table Schema for the \code{data}.
If not provided, one will be created using \code{\link[=create_schema]{create_schema()}}.}
}
\value{
Provided \code{package} with one additional resource.
}
\description{
Adds a Tabular \href{https://specs.frictionlessdata.io/data-resource/}{Data Resource} to a Data
Package.
The resource will be a \href{https://specs.frictionlessdata.io/tabular-data-resource/}{Tabular Data Resource}.
The resource name can only contain lowercase alphanumeric characters plus
\code{.}, \code{-} and \verb{_}.
}
\examples{
# Load the example Data Package
package <- example_package

# List the resource names
package$resource_names

# Create a data frame
df <- data.frame(
  multimedia_id = c(
    "aed5fa71-3ed4-4284-a6ba-3550d1a4de8d",
    "da81a501-8236-4cbd-aa95-4bc4b10a05df"
  ),
  x = c(718, 748),
  y = c(860, 900)
)

# Add a new resource "positions" to the Data Package, from the data frame
package <- add_resource(package, "positions", data = df)

# Add a new resource "deployments2" to the Data Package, from a CSV file path
csv_file_path <- system.file(
  "extdata", "deployments.csv", package = "frictionless"
)
package <- add_resource(package, "deployments2", data = csv_file_path)

# List the resource names ("positions" and "deployments2" added)
package$resource_names
}
\seealso{
Other edit functions: 
\code{\link{get_schema}()},
\code{\link{remove_resource}()}
}
\concept{edit functions}
