% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append.datapackage}
\alias{append.datapackage}
\title{Custom method for base append to retain custom S3 classes}
\usage{
\method{append}{datapackage}(x, values, after = length(x))
}
\arguments{
\item{x}{the vector the values are to be appended to.}

\item{values}{to be included in the modified vector.}

\item{after}{a subscript, after which the values are to be appended.}
}
\value{
Data Package
}
\description{
When base::append is used on a datapackage object, it loses it's custom s3
class. This is a custom method, that retains this s3 class.
}
\examples{
p <- create_package()
class(p)
p_appended <- append(p, c(new_record_to_append = "value to append"))
class(p_appended)
}
