% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append}
\alias{append}
\title{Append Method for Custom S3 Classes}
\usage{
\method{append}{default}(x, values, after = length(x))
}
\arguments{
\item{x}{The original object.}

\item{values}{The values to append to the object.}

\item{after}{A position after which to append the values.}
}
\value{
An object of the same class as x with values appended.
}
\description{
Append Method for Custom S3 Classes
}
