% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.list.R
\name{as.list.datapackage}
\alias{as.list.datapackage}
\title{Convert a Data Package to a list}
\usage{
\method{as.list}{datapackage}(package)
}
\value{
Data Package object converted to a list.
}
\description{
Converts a Data Package to a list, removing any custom classes including
\code{datapackage}.
To convert a list back to a Data Package, use \code{\link[=create_package]{create_package()}}.
}
\examples{
# Load the example Data Package
package <- example_package

# Convert the Data Package to a list
package_as_list <- as.list(package)

# Convert the list back to a Data Package
create_package(package_as_list)
}
\seealso{
Other datapackage functions: 
\code{\link{print.datapackage}()}
}
\concept{datapackage functions}
