% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_package.R
\name{check_package}
\alias{check_package}
\title{Check a Data Package object}
\usage{
check_package(package)
}
\arguments{
\item{package}{Data Package object, created with \code{\link[=read_package]{read_package()}} or
\code{\link[=create_package]{create_package()}}.}
}
\value{
\code{TRUE} or error.
}
\description{
Check if an object is a Data Package object with the required properties.
}
\concept{check functions}
