% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_package.R
\name{check_package}
\alias{check_package}
\title{Check package object}
\usage{
check_package(package)
}
\arguments{
\item{package}{Package object}
}
\value{
\code{TRUE} or error.
}
\description{
Check if a package object is a list object of class "datapackage".
}
\keyword{internal}
