% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_path.R
\name{check_path}
\alias{check_path}
\title{Check path or URL}
\usage{
check_path(path, directory = NULL, unsafe = TRUE)
}
\arguments{
\item{path}{Path or URL to a file.}

\item{directory}{Directory to prepend to path.}

\item{unsafe}{Allow \code{path} to be an unsafe absolute or relative parent path.}
}
\value{
Absolute path or URL.
}
\description{
Check if a
\href{https://specs.frictionlessdata.io/data-resource/#url-or-path}{path or URL}
is valid (and optionally safe) and prepend with directory to create an
absolute path or URL. Returns error when no file can be found.
}
\keyword{internal}
