% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_path}
\alias{check_path}
\title{Check URL or path}
\usage{
check_path(url_or_path, directory = NULL, unsafe = TRUE)
}
\arguments{
\item{url_or_path}{URL or path to a file.}

\item{directory}{Directory to prepend to path.}

\item{unsafe}{Allow unsafe absolute and relative parent paths.}
}
\value{
Full path or error.
}
\description{
Check if a
\href{https://specs.frictionlessdata.io/data-resource/#url-or-path}{URL or path}
is valid and prepend with directory to create a full path. Returns error when
no file can be found and optionally for unsafe absolute and relative parent
paths.
}
\keyword{internal}
