% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{create_package}
\alias{create_package}
\title{Create an empty Data Package}
\usage{
create_package()
}
\value{
List object describing a Data Package.
}
\description{
Initiates a list object describing a
\href{https://specs.frictionlessdata.io/data-package/}{Data Package}. This empty
Data Package can then be extended with metadata and resources (see
\code{\link[=add_resource]{add_resource()}}).
}
\examples{
# Create a Data Package
package <- create_package()
str(package)
}
\seealso{
Other create functions: 
\code{\link{create_schema}()}
}
\concept{create functions}
