% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{create_package}
\alias{create_package}
\title{Create a Data Package}
\usage{
create_package(descriptor = NULL)
}
\arguments{
\item{descriptor}{List to be made into a Data Package object.
If \code{NULL}, an empty Data Package object will be created from scratch.}
}
\value{
Data Package object.
Use \code{\link[=check_package]{check_package()}} to check if the output is a valid Data Package
object.
}
\description{
Initiates a \href{https://specs.frictionlessdata.io/data-package/}{Data Package}
object, either from scratch or from an existing list.
This Data Package object is a list with a \code{datapackage} class and the
following properties:
\itemize{
\item All properties of the original \code{descriptor}.
\item \href{https://specs.frictionlessdata.io/data-package/#required-properties}{\code{resources}}, if
not provided.
\item \href{https://specs.frictionlessdata.io/data-package/#profile}{\code{profile}}, if
not provided.
This property is set to \code{"tabular-data-package"}, since added resources
will make the Data Package meet \href{https://specs.frictionlessdata.io/tabular-data-package/}{Tabular Data Package} requirements.
\item \code{directory}, if not provided.
It is used as the base path to read resources with \code{\link[=read_resource]{read_resource()}}, with
the current directory (\code{"."}) by default.
}
}
\examples{
# Create a Data Package
package <- create_package()
str(package)
}
\seealso{
Other create functions: 
\code{\link{create_schema}()}
}
\concept{create functions}
