% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{create_package}
\alias{create_package}
\title{Create an empty Data Package}
\usage{
create_package()
}
\value{
List object of class \code{datapackage}.
}
\description{
Initiates a \href{https://specs.frictionlessdata.io/data-package/}{Data Package}
descriptor as a list object of class \code{datapackage} (cf. \code{\link[=read_package]{read_package()}}).
Can be extended with metadata and resources (see \emph{add_resource()}).
}
\examples{
# Create a Data Package
package <- create_package()
str(package)
}
