% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{create_package}
\alias{create_package}
\title{Create an empty Data Package}
\usage{
create_package(
  descriptor = list(profile = "tabular-data-package", resources = list(), directory =
    ".")
)
}
\arguments{
\item{descriptor}{List describing a data package}
}
\value{
A data package
}
\description{
Initiates a list describing a \href{https://specs.frictionlessdata.io/data-package/}{Data Package}.
This empty Data Package can be extended with metadata and resources (see
\code{\link[=add_resource]{add_resource()}}).
Added resources will make the Data Package meet \href{https://specs.frictionlessdata.io/tabular-data-package/}{Tabular Data Package}
requirements, so \code{profile} is set to \code{tabular-data-package}.
}
\examples{
# Create a Data Package
package <- create_package()
str(package)
}
\seealso{
Other create functions: 
\code{\link{create_schema}()}
}
\concept{create functions}
