% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_schema.R
\name{create_schema}
\alias{create_schema}
\title{Create a Table Schema for a data frame}
\usage{
create_schema(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
List object describing the Table Schema.
}
\description{
Creates a \href{https://specs.frictionlessdata.io/table-schema/}{Table Schema} for
a data frame. The schema will contain all data frame fields, their \code{name} and
\code{type}.
}
\examples{
# Create data frame
df <- data.frame(
  id = c(as.integer(1), as.integer(2)),
  timestamp = c(as.POSIXct("2020-03-01T12:00:00"), as.POSIXct("2020-03-01T18:45:00")),
  species = c("Capreolus capreolus", "Sus scrofa"),
  lifeStage = factor(c("adult", "adult"), levels = c("adult", "juvenile", "unknown"))
)

# Create Table Schema
schema <- create_schema(df)
str(schema)
}
