% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_schema.R
\name{create_schema}
\alias{create_schema}
\title{Create a Table Schema for a data frame}
\usage{
create_schema(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
List object with Table Schema properties.
}
\description{
Creates a \href{https://specs.frictionlessdata.io/table-schema/}{Table Schema} for
a data frame. \code{fields} will have properties \code{name} = column name and \code{type} =
translated column class. \code{missingValues} are set to \verb{c("", "NA"}).
}
\examples{
# Create data frame
df <- data.frame(
  id = c(as.integer(1), as.integer(2)),
  timestamp = c(as.POSIXct("2020-03-01T12:00:00"), as.POSIXct("2020-03-01T18:45:00")),
  species = c("Capreolus capreolus", "Sus scrofa")
)

# Create and print its schema
schema <- create_schema(df)
jsonlite::toJSON(schema, pretty = TRUE, auto_unbox = TRUE)
}
