% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_package.R
\name{example_package}
\alias{example_package}
\title{Read the example Data Package}
\usage{
example_package()
}
\value{
Data Package object, see \code{\link[=create_package]{create_package()}}.
}
\description{
Reads the example Data Package included in \code{frictionless}.
This dataset is used in examples, vignettes, and tests and contains dummy
camera trap data organized in 3 Data Resources:
\enumerate{
\item \code{deployments}: local data file referenced in
\code{"path": "deployments.csv"}.
\item \code{observations}: two remote data files referenced in
\verb{"path": ["url/to/observations_1.csv", "url/to/observations_2.csv"]}.
This resource requires an internet connection to be read.
\item \code{media}: inline data stored in \code{data}.
}
}
\examples{
example_package()
}
\concept{sample data}
