% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_package}
\alias{example_package}
\title{Example Data Package}
\format{
An object of class \code{datapackage} (inherits from \code{list}) of length 9.
}
\source{
\url{https://github.com/frictionlessdata/frictionless-r/tree/main/inst/extdata}
}
\usage{
example_package
}
\description{
Example Tabular
\href{https://specs.frictionlessdata.io/data-package/}{Data Package} with dummy
camera trap data organized in 3 Data Resources:
\itemize{
\item \code{deployments}: data stored in \code{deployments.csv}.
\item \code{observations}: data stored in \code{observations_1.csv} and
\code{observations_2.csv}.
\item \code{media}: data stored in \code{data} property.
}
}
\examples{
\dontrun{
# File was created with
example_package <- read_package(
  system.file("extdata", "datapackage.json", package = "frictionless")
)
save(example_package, file = "data/example_package")
}
}
\keyword{datasets}
