% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_package.R
\name{example_package}
\alias{example_package}
\title{Read the example Data Package}
\usage{
example_package(version = "1.0")
}
\arguments{
\item{version}{Version of the \href{https://datapackage.org/blog/2024-06-26-v2-release/}{Data Package} standard that this
example Data Package uses. Currently only version 1.0 is supported.}
}
\value{
A Data Package object, see \code{\link[=create_package]{create_package()}}.
}
\description{
Reads the example Data Package included in \code{frictionless}.
This dataset is used in examples, vignettes, and tests and contains dummy
camera trap data organized in 3 Data Resources:
\enumerate{
\item \code{deployments}: one local data file referenced in
\code{"path": "deployments.csv"}.
\item \code{observations}: two local data files referenced in
\verb{"path": ["observations_1.tsv", "observations_2.tsv"]}.
\item \code{media}: inline data stored in \code{data}.
}
}
\examples{
example_package()
}
\concept{sample data}
