% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frictionless-package.R
\docType{package}
\name{frictionless-package}
\alias{frictionless}
\alias{frictionless-package}
\title{frictionless: Read and Write Frictionless Data Packages}
\description{
Read and write Frictionless Data Packages. A 'Data Package' (\url{https://specs.frictionlessdata.io/data-package/}) is a simple container format and standard to describe and package a collection of (tabular) data. It is typically used to publish FAIR (\url{https://www.go-fair.org/fair-principles/}) and open datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frictionlessdata/frictionless-r}
  \item \url{https://docs.ropensci.org/frictionless/}
  \item Report bugs at \url{https://github.com/frictionlessdata/frictionless-r/issues}
}

}
\author{
\strong{Maintainer}: Peter Desmet \email{peter.desmet.work@gmail.com} (\href{https://orcid.org/0000-0002-8442-8025}{ORCID})

Authors:
\itemize{
  \item Damiano Oldoni \email{damiano.oldoni@inbo.be} (\href{https://orcid.org/0000-0003-3445-7562}{ORCID})
}

Other contributors:
\itemize{
  \item Research Institute for Nature and Forest (INBO) (https://www.vlaanderen.be/inbo/en-gb/) [copyright holder]
  \item LifeWatch Belgium (https://lifewatch.be) [funder]
  \item Beatriz Milz \email{milz.bea@gmail.com} (\href{https://orcid.org/0000-0002-3064-4486}{ORCID}) [reviewer]
  \item João Martins \email{joao.martins@gmx.net} (\href{https://orcid.org/0000-0001-7961-4280}{ORCID}) [reviewer]
}

}
\keyword{internal}
