% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frictionless.R
\docType{package}
\name{frictionless-package}
\alias{frictionless}
\alias{frictionless-package}
\title{frictionless: Read and Write Frictionless Data Packages}
\description{
Read and write Frictionless Data Packages in a Tidyverse way.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frictionlessdata/frictionless-r}
  \item \url{https://frictionlessdata.github.io/frictionless-r}
  \item Report bugs at \url{https://github.com/frictionlessdata/frictionless-r/issues}
}

}
\author{
\strong{Maintainer}: Peter Desmet \email{peter.desmet@inbo.be} (\href{https://orcid.org/0000-0002-8442-8025}{ORCID})

Authors:
\itemize{
  \item Damiano Oldoni \email{damiano.oldoni@inbo.be} (\href{https://orcid.org/0000-0003-3445-7562}{ORCID})
}

}
\keyword{internal}
