% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resource.R
\name{get_resource}
\alias{get_resource}
\title{Get a Data Resource}
\usage{
get_resource(resource_name, package)
}
\arguments{
\item{resource_name}{Name of the resource.}

\item{package}{List object describing a Data Package, created with
\code{\link[=read_package]{read_package()}} or \code{\link[=create_package]{create_package()}}.}
}
\value{
List object describing a Data Resource, with new property \code{full_path}
containing absolute \code{path}s used for reading data.
}
\description{
Returns a \href{https://specs.frictionlessdata.io/data-resource/}{Data Resource}
from a Data Package, i.e. the content of one of the described \code{resources}.
}
\examples{
# Read a datapackage.json file
package <- read_package(
  system.file("extdata", "datapackage.json", package = "frictionless")
)

# Get the Data Resource "observations"
resource <- get_resource("observations", package)
str(resource)
}
