% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resource.R
\name{get_resource}
\alias{get_resource}
\title{Get a Data Resource}
\usage{
get_resource(package, resource_name)
}
\arguments{
\item{package}{List object describing a Data Package, created with
\code{\link[=read_package]{read_package()}} or \code{\link[=create_package]{create_package()}}.}

\item{resource_name}{Name of the resource.}
}
\value{
List object describing a Data Resource, with new property \code{read_from}
to indicate how data should be read. If present, \code{path} will be updated to
contain the full path(s).
}
\description{
Returns a \href{https://specs.frictionlessdata.io/data-resource/}{Data Resource}
from a Data Package, i.e. the content of one of the described \code{resources}.
}
\examples{
# Read a datapackage.json file
package <- read_package(
  system.file("extdata", "datapackage.json", package = "frictionless")
)

# Get the Data Resource "observations"
resource <- get_resource(package, "observations")
str(resource)
}
