% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resource.R
\name{get_resource}
\alias{get_resource}
\title{Get a Data Package resource}
\usage{
get_resource(resource_name, package)
}
\arguments{
\item{resource_name}{Name of the resource.}

\item{package}{Data Package object, see \code{read_package()}.}
}
\value{
List object describing a Data Resource.
}
\description{
Returns a \href{https://specs.frictionlessdata.io/data-resource/}{Data Resource}
from a Data Package, i.e. the content of one of the described \code{resources}.
}
\examples{
# Read datapackage.json file
package <- read_package(system.file("extdata", "datapackage.json", package = "datapackage"))

# Get resource "observations"
resource <- get_resource("observations", package)
str(resource)
}
