% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_schema.R
\name{get_schema}
\alias{get_schema}
\title{Get the Table Schema of a Data Resource}
\usage{
get_schema(resource_name, package)
}
\arguments{
\item{resource_name}{Name of the resource.}

\item{package}{List object describing a Data Package, created with
\code{\link[=read_package]{read_package()}} or \code{\link[=create_package]{create_package()}}.}
}
\value{
List object describing a Table Schema.
}
\description{
Returns the \href{https://specs.frictionlessdata.io/table-schema/}{Table Schema}
of a Data Resource (in a Data Package), i.e. the content of its \code{schema}
property, describing the resource's fields, data types, relationships, and
missing values. The resource must be a
\href{https://specs.frictionlessdata.io/tabular-data-resource/}{Tabular Data Resource}.
}
\examples{
# Read a datapackage.json file
package <- read_package(
  system.file("extdata", "datapackage.json", package = "frictionless")
)

# Get the Table Schema for resource "observations"
schema <- get_schema("observations", package)
str(schema)
}
