% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.datapackage}
\alias{print.datapackage}
\title{Print a data package}
\usage{
\method{print}{datapackage}(x)
}
\arguments{
\item{x}{Data package to print}
}
\description{
Print a data package
}
\examples{
# Read a datapackage.json file
package <- read_package(
  system.file("extdata", "datapackage.json", package = "frictionless")
)

# Print out a summary of the data package
print(package)
}
\concept{print functions}
