% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.datapackage}
\alias{print.datapackage}
\title{Print a Data Package}
\usage{
\method{print}{datapackage}(x, ...)
}
\arguments{
\item{x}{Data Package object, created with \code{\link[=read_package]{read_package()}} or
\code{\link[=create_package]{create_package()}}.}

\item{...}{Further arguments, they are ignored by this function.}
}
\value{
\code{\link[=print]{print()}} with a summary of the Data Package object.
}
\description{
Prints a human-readable summary of a Data Package, including its resources
and a link to more information (if provided in \code{package$id}).
}
\examples{
# Load the example Data Package
package <- example_package()

# Print a summary of the Data Package
package # Or print(package)
}
\concept{datapackage functions}
