% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-problems.R
\name{problems}
\alias{problems}
\title{Retrieve parsing problems}
\description{
Allows to show parsing issues in data frame returned by \code{\link[=read_resource]{read_resource()}}.
See \code{\link[readr:problems]{readr::problems()}} for details.
}
\keyword{internal}
