% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_descriptor.R
\name{read_descriptor}
\alias{read_descriptor}
\title{Read a datapackage.json file}
\usage{
read_descriptor(descriptor_file = "datapackage.json")
}
\arguments{
\item{descriptor_file}{Path to a \code{datapackage.json} file.}
}
\value{
List object containing descriptor information and two new properties:
\itemize{
\item \code{resource_names}: vector with resource names.
\item \code{directory}: directory path of descriptor file, used as base path to read
resources with \code{read_resource()}.
}
}
\description{
Reads information from a \code{datapackage.json} file, i.e. the
\href{https://specs.frictionlessdata.io/data-package/#descriptor}{descriptor}
that describes the Data Package metadata and its resources.
}
\examples{
descriptor <- read_descriptor(system.file("extdata", "datapackage.json", package = "datapackage"))
descriptor$resource_names
}
