% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_package.R
\name{read_package}
\alias{read_package}
\title{Read a datapackage.json file}
\usage{
read_package(descriptor_file = "datapackage.json")
}
\arguments{
\item{descriptor_file}{Path to a \code{datapackage.json} file.}
}
\value{
List object containing the descriptor information and two new
properties:
\itemize{
\item \code{resource_names}: vector with resource names.
\item \code{directory}: directory path of descriptor file, used as base path to read
resources with \code{read_resource()}.
}
}
\description{
Reads information from a \code{datapackage.json} file, i.e. the
\href{https://specs.frictionlessdata.io/data-package/#descriptor}{descriptor} that
describes the Data Package metadata and its resources.
}
\examples{
path <- system.file("extdata", "datapackage.json", package = "datapackage")
package <- read_package(path)
package$resource_names
}
