% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_package.R
\name{read_package}
\alias{read_package}
\title{Read a Data Package descriptor file (\code{datapackage.json})}
\usage{
read_package(file = "datapackage.json")
}
\arguments{
\item{file}{Path or URL to a \code{datapackage.json} file.}
}
\value{
List object describing a Data Package, i.e the content of the
descriptor file and two new properties:
\itemize{
\item \code{resource_names}: vector with resource names.
\item \code{directory}: path to Data Package directory, used as base path to read
resources with \code{\link[=read_resource]{read_resource()}}.
}
}
\description{
Reads information from a \code{datapackage.json} file, i.e. the
\href{https://specs.frictionlessdata.io/data-package/#descriptor}{descriptor} file
that describes the Data Package metadata and its resources.
}
\examples{
# Read a datapackage.json file
package <- read_package(
  system.file("extdata", "datapackage.json", package = "frictionless")
)

# Access the package properties
package$name
package$resource_names
}
\seealso{
Other read functions: 
\code{\link{read_resource}()}
}
\concept{read functions}
