% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_package.R
\name{read_package}
\alias{read_package}
\title{Read a Data Package descriptor file (\code{datapackage.json})}
\usage{
read_package(file = "datapackage.json")
}
\arguments{
\item{file}{Path or URL to a \code{datapackage.json} file.}
}
\value{
List object of class \code{datapackage}, containing the descriptor
information and two new properties:
\itemize{
\item \code{resource_names}: vector with resource names.
\item \code{directory}: path to Data Package directory, used as root path to read
resources with \code{read_resource()}.
}
}
\description{
Reads information from a \code{datapackage.json} file, i.e. the
\href{https://specs.frictionlessdata.io/data-package/#descriptor}{descriptor} that
describes the Data Package metadata and its resources.
}
\examples{
# Read datapackage.json file
package <- read_package(system.file("extdata", "datapackage.json", package = "frictionless"))

# Access package properties
package$name
package$resource_names
}
