% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_resource.R
\name{read_resource}
\alias{read_resource}
\title{Read data from a Data Package resource}
\usage{
read_resource(descriptor, resource_name)
}
\arguments{
\item{descriptor}{Descriptor object (see \code{read_descriptor()}).}

\item{resource_name}{Name of the resource to load data from.}
}
\value{
tibble with the resource data.
}
\description{
Loads data from a Data Package \strong{resource} into a tibble (a Tidyverse
data.frame).
}
\examples{
\dontrun{
descriptor <- read_descriptor(system.file("extdata", "datapackage.json", package = "datapackage"))
df <- read_resource(descriptor, "deployments")
}
}
