% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_resource.R
\name{remove_resource}
\alias{remove_resource}
\title{Remove a Data Resource}
\usage{
remove_resource(resource_name, package)
}
\arguments{
\item{resource_name}{Name of the resource.}

\item{package}{Data Package object, see \code{read_package()}.}
}
\value{
Data Package object.
}
\description{
Removes a \href{https://specs.frictionlessdata.io/data-resource/}{Data Resource}
from a Data Package, i.e. one of the described \code{resources} is removed.
}
\examples{
# Read datapackage.json file
package <- read_package(system.file("extdata", "datapackage.json", package = "datapackage"))

# List resource names
package$resource_names

# Remove resource "observations"
package <- remove_resource("observations", package)

# List resource names
package$resource_names
}
