% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_resource.R
\name{remove_resource}
\alias{remove_resource}
\title{Remove a Data Resource}
\usage{
remove_resource(package, resource_name)
}
\arguments{
\item{package}{List object describing a Data Package, created with
\code{\link[=read_package]{read_package()}} or \code{\link[=create_package]{create_package()}}.}

\item{resource_name}{Name of the resource.}
}
\value{
List object describing a Data Package.
}
\description{
Removes a \href{https://specs.frictionlessdata.io/data-resource/}{Data Resource}
from a Data Package, i.e. one of the described \code{resources} is removed.
}
\examples{
# Read a datapackage.json file
package <- read_package(system.file("extdata", "datapackage.json", package = "frictionless"))

# List the resource names
package$resource_names

# Remove the resource "observations"
package <- remove_resource(package, "observations")

# List the resource names ("observations" removed)
package$resource_names
}
