% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{resource}
\alias{resource}
\alias{resource<-}
\title{Get or overwrite a Data Resource}
\usage{
resource(package, resource_name)

resource(package, resource_name) <- value
}
\arguments{
\item{package}{Data Package object, as returned by \code{\link[=read_package]{read_package()}} or
\code{\link[=create_package]{create_package()}}.}

\item{resource_name}{Name of the Data Resource.}

\item{value}{Value to assign.}
}
\value{
List describing a Data Resource.

\code{package} with overwritten resource.
}
\description{
Gets or overwrites a Data Resource from/in a Data Package.
These functions are \strong{designed for internal use in other packages}.
For public manipulation of resources, use \code{\link[=add_resource]{add_resource()}} and
\code{\link[=remove_resource]{remove_resource()}}.

\code{resource()} gets a Data Resource from a Data Package by its name.
The returned value will be a list describing a Data Resource, with a new
attribute \code{data_location} to indicate how the data are attached.
If present, \code{path} will be updated to the full path(s).

\verb{resource<-} overwrites a Data Resource in Data Package with a new value.
The assigned value will typically be obtained with \code{resource()} and then
manipulated, so the function will revert internal changes made by
\code{resource()} (i.e. removing the attribute \code{data_location} and updating paths
to the original values).
The assigned value is otherwise \strong{not validated}, so use with care.
}
\details{
See \code{vignette("data-resource")} to learn more about Data Resource.
}
\examples{
# Load the example Data Package
package <- example_package()

# Get the resource "deployments"
resource <- resource(package, "deployments")
str(resource)

# Update the resource
resource$title <- "Camera trap deployments"

# Overwrite the resource
resource(package, "deployments") <- resource
}
\seealso{
Other accessor functions:
\code{\link[=resource_names]{resource_names()}},
\code{\link[=schema]{schema()}}
}
\concept{accessor functions}
