% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{resource}
\alias{resource}
\alias{resource<-}
\title{Get or set a Data Resource}
\usage{
resource(package, resource_name)

resource(package, resource_name) <- value
}
\arguments{
\item{package}{Data Package object, as returned by \code{\link[=read_package]{read_package()}} or
\code{\link[=create_package]{create_package()}}.}

\item{resource_name}{Name of the Data Resource.}

\item{value}{Data Resource to assign.}
}
\value{
List describing a Data Resource.

\code{package} with assigned resource.
}
\description{
\code{resource()} gets a Data Resource from a Data Package by its name.
The returned value will be a list describing a Data Resource, with a new
attribute \code{data_location} to indicate how the data are attached.
If present, \code{path} will be updated to contain the full path(s).

\verb{resource<-} assigns a Data Resource to a Data Package by its name.
If present, the assigned value will be stripped from its \code{data_location}
attribute and \code{path} shortened to its original value (without the prepended
\code{package$directory}).
The value is otherwise not validated.

Note that these functions are designed for internal use in other packages.
For public manipulation of resources, use \code{\link[=add_resource]{add_resource()}} and
\code{\link[=remove_resource]{remove_resource()}}.
}
\examples{
# Load the example Data Package
package <- example_package()

# Get the resource "deployments"
resource <- resource(package, "deployments")
str(resource)

# Set the resource "deployments"
resource(package, "deployments") <- resource
}
\seealso{
Other accessor functions:
\code{\link[=resource_names]{resource_names()}},
\code{\link[=schema]{schema()}}
}
\concept{accessor functions}
