% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_names.R
\name{resource_names}
\alias{resource_names}
\title{List Data Resource names}
\usage{
resource_names(package)
}
\arguments{
\item{package}{Data Package object, as returned by \code{\link[=read_package]{read_package()}} or
\code{\link[=create_package]{create_package()}}.}
}
\value{
Character vector with the Data Resource names.
}
\description{
Lists the names of the Data Resources included in a Data Package.
}
\examples{
# Load the example Data Package
package <- example_package()

# List the resources
resource_names(package)
}
\seealso{
Other accessor functions: 
\code{\link{schema}()}
}
\concept{accessor functions}
