% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{resources}
\alias{resources}
\title{List Data Resources}
\usage{
resources(package)
}
\arguments{
\item{package}{List describing a Data Package.}
}
\value{
Character vector with the resource names.
}
\description{
List the names of the Data Resources included in a Data Package
}
\examples{
# Load the example Data Package
package <- example_package

# List the resource names
resources(package)
}
\seealso{
Other read functions: 
\code{\link{read_package}()},
\code{\link{read_resource}()}
}
\concept{read functions}
