% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{resources}
\alias{resources}
\title{List Data Resources}
\usage{
resources(package)
}
\arguments{
\item{package}{Data Package object, as returned by \code{\link[=read_package]{read_package()}} or
\code{\link[=create_package]{create_package()}}.}
}
\value{
Character vector with the Data Resource names.
}
\description{
Lists the names of the Data Resources included in a Data Package.
}
\examples{
# Load the example Data Package
package <- example_package()

# List the resources
resources(package)
}
\seealso{
Other read functions: 
\code{\link{read_package}()},
\code{\link{read_resource}()}
}
\concept{read functions}
