% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_package.R
\name{upgrade_package}
\alias{upgrade_package}
\title{Upgrade a Data Package to v2}
\usage{
upgrade_package(package)
}
\arguments{
\item{package}{Data Package object, as returned by \code{\link[=read_package]{read_package()}} or
\code{\link[=create_package]{create_package()}}.}
}
\value{
Upgraded \code{package}.
}
\description{
Upgrades a Data Package, its Data Resources and Table Schemas to the
\href{https://datapackage.org/}{v2} specification.
}
\examples{
# Load the v1 example Data Package
(package <- example_package(version = "1.0"))

# Upgrade
(package_upgraded <- upgrade_package(package))

# Check version of one of the resources
version(resource(package_upgraded, "deployments"))
}
\seealso{
Other versioning functions:
\code{\link[=version]{version()}}
}
\concept{versioning functions}
