% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_package.R
\name{upgrade_package}
\alias{upgrade_package}
\title{Upgrade a Data Package to v2}
\usage{
upgrade_package(package)
}
\arguments{
\item{package}{Data Package object, as returned by \code{\link[=read_package]{read_package()}} or
\code{\link[=create_package]{create_package()}}.}
}
\value{
Upgraded \code{package}.
}
\description{
Upgrades a Data Package, its Data Resources and Table Schemas to the
\href{https://datapackage.org/}{v2} specification.
}
\section{Upgrade details}{

\subsection{Data Package}{

\code{upgrade_package()} upgrades a \href{https://specs.frictionlessdata.io/data-package/}{v1} descriptor to \href{https://datapackage.org/standard/data-package/}{v2} as follows:
\itemize{
\item Adds \verb{$schema} as first property and sets it to the recommended v2 value
(\code{"https://datapackage.org/profiles/2.0/datapackage.json"}), except for
certain \code{profile} values.
\item Removes \code{profile}, but retains its value in \verb{$schema} if it is a URL to a
custom profile (see \href{https://datapackage.org/standard/data-package/#dollar-schema}{backwards compatibility}).
\item Converts \code{contributors} \code{"role": "value"} to \verb{"roles": ["value"]} (see
\href{https://datapackage.org/overview/changelog/#packagecontributors-updated}{backwards compatibility}).
}
}

\subsection{Data Resource}{

\code{upgrade_package()} upgrades any \href{https://specs.frictionlessdata.io/data-resource/}{v1} resource to \href{https://datapackage.org/standard/data-resource/}{v2} as follows:
\itemize{
\item Adds \verb{$schema} as first property and sets it to the recommended v2 value
(\code{"https://datapackage.org/profiles/2.0/dataresource.json"}).
\item Removes \code{profile}, but converts \code{"profile" = "tabular-data-resource"} to
\code{"type" = "table"} (see \href{https://datapackage.org/standard/data-resource/#type}{backwards compatibility}).
}
}

\subsection{Table Dialect}{

\code{upgrade_package()} leaves the dialect as is for all resources.
}

\subsection{Table Schema}{

\code{upgrade_package()} upgrades any \href{https://specs.frictionlessdata.io/table-schema/}{v1} schema to \href{https://datapackage.org/standard/table-schema/}{v2} as follows:
\itemize{
\item Adds \verb{$schema} as first property and sets it to the recommended v2 value
(\code{"https://datapackage.org/profiles/2.0/tableschema.json"}).
\item Converts \code{primaryKey} single values to an array (see \href{https://datapackage.org/overview/changelog/#schemaprimarykey-updated}{backwards compatibility}).
\item Converts \code{foreignKeys} single values in \code{fields} to an array and removes
\code{reference$resource} if it is self-referential (see \href{https://datapackage.org/overview/changelog/#schemaprimarykey-updated}{backwards compatibility}).
}
}
}

\examples{
# Load the v1 example Data Package
(package <- example_package(version = "1.0"))

# Upgrade
(package_upgraded <- upgrade_package(package))

# Check version of one of the resources
version(resource(package_upgraded, "deployments"))
}
\seealso{
Other versioning functions:
\code{\link[=version]{version()}}
}
\concept{versioning functions}
