% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{version}
\alias{version}
\title{Get Data Package version}
\usage{
version(package)
}
\arguments{
\item{package}{Data Package object, as returned by \code{\link[=read_package]{read_package()}} or
\code{\link[=create_package]{create_package()}}.}
}
\value{
Data Package version number (e.g. \code{1.0}).
}
\description{
Determines what version of the \href{https://datapackage.org/}{Data Package standard} is used by a Data Package, based on the
\href{https://datapackage.org/standard/data-package/#dollar-schema}{\verb{$schema}}
property.
Version \code{"1.0"} is assumed if \verb{$schema} is missing, version \code{">=2.0"} is
assumed for custom values (e.g. \href{https://datapackage.org/standard/extensions/}{extensions}).
}
\examples{
package <- example_dataset()
version(package)
}
\concept{version functions}
