% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{version}
\alias{version}
\title{Get Data Package version}
\usage{
version(x)
}
\arguments{
\item{x}{A list describing either a:
\itemize{
\item Data Package.
\item Data Resource.
\item Table Dialect.
\item Table Schema.
}}
}
\value{
Data Package standard version number (e.g. \code{"1.0"}).
}
\description{
Determines what version of the \href{https://datapackage.org/}{Data Package standard} is used by a Data Package, Data Resource, Table
Dialect or Table Schema, based on the \href{https://datapackage.org/standard/data-package/#dollar-schema}{\verb{$schema}} property.
Version \code{"1.0"} is assumed if \verb{$schema} is missing, version \code{">=2.0"} is
assumed for custom values (e.g. \href{https://datapackage.org/standard/extensions/}{extensions}).
}
\examples{
# Data Package
package <- example_package()
version(package)

# Data Resource
resource <- resource(package, "observations")
version(resource)

# Table Dialect
version(resource$dialect)

# Table Schema
schema <- schema(package, "observations")
version(schema)
}
\concept{version functions}
