% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_package.R
\name{write_package}
\alias{write_package}
\title{Write a Data Package to disk}
\usage{
write_package(package, directory, compress = FALSE)
}
\arguments{
\item{package}{Data Package object, as returned by \code{\link[=read_package]{read_package()}} or
\code{\link[=create_package]{create_package()}}.}

\item{directory}{Path to local directory to write files to.}

\item{compress}{If \code{TRUE}, data of added resources will be gzip compressed
before being written to disk (e.g. \code{deployments.csv.gz}).}
}
\value{
\code{package} invisibly, as written to file.
}
\description{
Writes a Data Package and its related Data Resources to disk as a
\code{datapackage.json} and CSV files.
}
\section{Writing data to CSV files}{


\code{write_package()} will write data to CSV files depending on how these are
attached to the Data Resource:
\itemize{
\item \strong{Data frame}:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{add_resource(package, "media", data = df)
}\if{html}{\out{</div>}}

Data are written to a CSV file in \code{directory} using \code{\link[readr:write_csv]{readr::write_csv()}}.
The CSV file will have the same name as the resource, overwriting any
existing file with the same name.
Use \code{compress = TRUE} to gzip the CSV file.
\item One or more \strong{paths} to local CSV files in a \strong{different directory}:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{add_resource(package, "media", data = "other-directory/media.csv")
}\if{html}{\out{</div>}}

CSV files are copied to \code{directory}, overwriting existing files with
the same name.
\item One or more \strong{paths} to local CSV files in the \strong{same directory}:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{add_resource(package, "media", data = "directory/media.csv")
}\if{html}{\out{</div>}}

CSV files are left as is (no overwrite).
This allows you to read and write a \code{datapackage.json} to the same
directory, without altering the CSV files of resources you did not
manipulate.
\item One or more \strong{URLs} to CSV files:

\if{html}{\out{<div class="sourceCode">}}\preformatted{add_resource(package, "media", data = "https://example.org/media.csv")
}\if{html}{\out{</div>}}

Files are not downloaded.
\item Mix of \strong{URLs and paths} to CSV files:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{add_resource(
  package, "media",
  data = c("https://example.org/media.csv", "media.csv")
)
}\if{html}{\out{</div>}}

Remote CSV files are downloaded to \code{directory}, overwriting existing
files with the same name.
Local CSV files are handled as described above.
\item \strong{Inline data}: No files are written.
}

In all above cases \code{path} is added or updated to the new file location(s)
when appropriate.
}

\examples{
# Load the example Data Package from disk
package <- read_package(
  system.file("extdata", "v1", "datapackage.json", package = "frictionless")
)

package

# Write the (unchanged) Data Package to disk
write_package(package, directory = "my_directory")

# Check files
list.files("my_directory")

# No files written for the "observations" resource, since those are all URLs.
# No files written for the "media" resource, since it has inline data.

# Clean up (don't do this if you want to keep your files)
unlink("my_directory", recursive = TRUE)
}
\concept{write functions}
