% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_package.R
\name{write_package}
\alias{write_package}
\title{Write a Data Package to disk}
\usage{
write_package(package, directory, compress = FALSE)
}
\arguments{
\item{package}{Data Package object, as returned by \code{\link[=read_package]{read_package()}} or
\code{\link[=create_package]{create_package()}}.}

\item{directory}{Path to local directory to write files to.}

\item{compress}{If \code{TRUE}, data of added resources will be gzip compressed
before being written to disk (e.g. \code{deployments.csv.gz}).}
}
\value{
\code{package} invisibly, as written to file.
}
\description{
Writes a Data Package and its related Data Resources to disk as a
\code{datapackage.json} file.
For some resources, it also writes the data as CSV files.
See \code{vignette("data-resource")} for details.
}
\examples{
# Load the example Data Package from disk
package <- read_package(
  system.file("extdata", "v1", "datapackage.json", package = "frictionless")
)

package

# Write the (unchanged) Data Package to disk
write_package(package, directory = "my_directory")

# Check files
list.files("my_directory")

# No files written for the "observations" resource, since those are all URLs.
# No files written for the "media" resource, since it has inline data.

# Clean up (don't do this if you want to keep your files)
unlink("my_directory", recursive = TRUE)
}
\concept{write functions}
