% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_package.R
\name{write_package}
\alias{write_package}
\title{Write a Data Package to disk}
\usage{
write_package(package, directory, compress = FALSE)
}
\arguments{
\item{package}{Data Package object, as returned by \code{\link[=read_package]{read_package()}} or
\code{\link[=create_package]{create_package()}}.}

\item{directory}{Path to local directory to write files to.}

\item{compress}{If \code{TRUE}, data of added resources will be gzip compressed
before being written to disk (e.g. \code{deployments.csv.gz}).}
}
\value{
\code{package} invisibly, as written to file.
}
\description{
Writes a Data Package and its related Data Resources to disk as a
\code{datapackage.json}.
Depending how the data are attached to the Data Resource, the function will
also write or copy these to CSV files.
}
\section{Writing data to CSV files}{


For Data Resources added or replaced with \code{add_resource()}, with data
attached as:
\itemize{
\item A data frame: data are written to a CSV file in \code{directory} using
\code{readr::write_csv()} and \code{path} is added.
The CSV file will have the same name as the resource and can be compressed
with \code{compress = TRUE}.
\item One or more paths to local CSV file(s): CSV file(s) are copied to
\code{directory} and \code{path} is updated to the new location.
\item URL(s) to CSV file(s): no CSV files are written.
}

For existing Data Resources that were not manipulated, with data attached as:
\itemize{
\item One or more paths to local CSV file(s): CSV file(s) are copied to
\code{directory} and \code{path} is updated to the new location.
Existing CSV files of the same name in \code{directory} are \emph{not overwritten}.
This allows you to read and write a \code{datapackage.json} to the same
directory, without altering the files of resources you did not manipulate.
\item URL(s) to CSV file(s): no CSV files are written.
\item Inline data: no CSV files are written.
}
}

\examples{
# Load the example Data Package from disk
package <- read_package(
  system.file("extdata", "v1", "datapackage.json", package = "frictionless")
)

package

# Write the (unchanged) Data Package to disk
write_package(package, directory = "my_directory")

# Check files
list.files("my_directory")

# No files written for the "observations" resource, since those are all URLs.
# No files written for the "media" resource, since it has inline data.

# Clean up (don't do this if you want to keep your files)
unlink("my_directory", recursive = TRUE)
}
\concept{write functions}
