% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_package.R
\name{write_package}
\alias{write_package}
\title{Write a Data Package to disk}
\usage{
write_package(package, directory = ".")
}
\arguments{
\item{package}{List object describing a Data Package, created with
\code{\link[=read_package]{read_package()}} or \code{\link[=create_package]{create_package()}}.}

\item{directory}{Path to local directory to write files to.}
}
\value{
Provided \code{package} (invisibly).
}
\description{
Writes a Data Package and its related Data Resources to disk as a
\code{datapackage.json} and CSV files.
Data Resources are handled as follows:
\itemize{
\item Resource has URL \code{path}: no action.
\item Resource has a local \code{path}: CSV files are copied (not read) to \code{directory}
to avoid relative/absolute local paths.
\item Resource has inline \code{data}:
\itemize{
\item Originally included as such: no action.
\item As result of adding a data frame with \code{add_resource()}: data are written
to a CSV file using \code{\link[readr:write_delim]{readr::write_csv()}}; \code{data} property is removed.
}
}
}
\examples{
# Load the example Data Package from disk
package <- read_package(
  system.file("extdata", "datapackage.json", package = "frictionless")
)

# List the resource names
package$resource_names

# Write the (unchanged) Data Package to disk
write_package(package, directory = "my_package")

# Check files
list.files("my_package")

# Clean up (don't do this if you want to keep your files)
unlink("my_package", recursive = TRUE)
}
\concept{write functions}
